move easy challenge to separate source files
