+++ Challenge 380 smooshed morse code/Challenge 380 smooshed morse code.vcxproj.filters
@@ -18,5 +18,16 @@
     <ClCompile Include="smoosh.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="smorseeasy.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="smorseeasy.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="alphabet.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
