+++ Challenge 380 smooshed morse code/alphabet.h
@@ -1,0 +1,10 @@
+#pragma once
+
+const int alphabet_len = 26;
+const char ascii_alphabet[] = {'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i',
+                               'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r',
+                               's', 't', 'u', 'v', 'w', 'x', 'y', 'z'};
+const std::string morse_alphabet[] = {
+    ".-",   "-...", "-.-.", "-..",  ".",   "..-.", "--.",  "....", "..",
+    ".---", "-.-",  ".-..", "--",   "-.",  "---",  ".--.", "--.-", ".-.",
+    "...",  "-",    "..-",  "...-", ".--", "-..-", "-.--", "--.."};
